+++ hooks/use-session-sync.ts
@@ -8,7 +8,7 @@ import {
   resetAuthState,
   setDomain,
   setTenantId,
-} from "@/lib/redux/slices/app-config";
+} from "@/lib/redux/slices/app-state";
 
 export function useSessionSync() {
   const { data: session, status, update } = useSession();
@@ -16,7 +16,7 @@ export function useSessionSync() {
   const dispatch = useAppDispatch();
   const { handleError } = useErrorHandler();
   const lastCheckRef = useRef<number>(Date.now());
-  const appConfig = useAppSelector((state) => state.appConfig);
+  const appConfig = useAppSelector((state) => state.app);
 
   useEffect(() => {
     const checkInterval = setInterval(async () => {
